Move calibration draw tools & make them all scalable
