+++ frontend/src/services/canvasDrawTools.tsx
@@ -1,0 +1,91 @@
+import { RoomCalibrationPoint } from "./roomCalibration";
+
+export const maxCord = 640;
+
+export const mapCoordinate = (cord: number, canvasSize: number): number => {
+  // Remove aliasing using +0.5
+  return Math.round((cord / maxCord) * canvasSize) + 0.5;
+}
+
+export const getRandomHexColourString = (index: number): string => {
+  const maxAmountOfColours = 16;
+  const hueDelta = Math.trunc(360 / maxAmountOfColours);
+  const h = (index % maxAmountOfColours) * hueDelta;
+  let s = 100;
+  let l = 40;
+
+  // convert hsl to rgb from: https://css-tricks.com/converting-color-spaces-in-javascript/
+  s /= 100;
+  l /= 100;
+
+  let c = (1 - Math.abs(2 * l - 1)) * s,
+    x = c * (1 - Math.abs((h / 60) % 2 - 1)),
+    m = l - c / 2,
+    r = 0,
+    g = 0,
+    b = 0;
+
+  if (0 <= h && h < 60) {
+    r = c; g = x; b = 0;
+  } else if (60 <= h && h < 120) {
+    r = x; g = c; b = 0;
+  } else if (120 <= h && h < 180) {
+    r = 0; g = c; b = x;
+  } else if (180 <= h && h < 240) {
+    r = 0; g = x; b = c;
+  } else if (240 <= h && h < 300) {
+    r = x; g = 0; b = c;
+  } else if (300 <= h && h < 360) {
+    r = c; g = 0; b = x;
+  }
+  // Having obtained RGB, convert channels to hex
+  let rString = Math.round((r + m) * 255).toString(16);
+  let gString = Math.round((g + m) * 255).toString(16);
+  let bString = Math.round((b + m) * 255).toString(16);
+
+  // Prepend 0s, if necessary
+  if (rString.length === 1)
+    rString = "0" + r;
+  if (gString.length === 1)
+    gString = "0" + g;
+  if (bString.length === 1)
+    bString = "0" + b;
+
+  return "#" + rString + gString + bString;
+}
+
+export const drawCurrentPosition = (context: CanvasRenderingContext2D, canvasSize: number, x: number, y: number) => {
+  const crosshairRadius = 0.03 * canvasSize;
+  context.strokeStyle = '#ff0000';
+  context.lineWidth = 0.01 * canvasSize;
+  context.beginPath();
+  context.moveTo(x - crosshairRadius, y);
+  context.lineTo(x + crosshairRadius, y);
+  context.moveTo(x, y - crosshairRadius);
+  context.lineTo(x, y + crosshairRadius);
+  context.stroke();
+}
+
+export const drawPoints = (context: CanvasRenderingContext2D, canvasSize: number, previousPoints: RoomCalibrationPoint[], fillStyle: string) => {
+  const pointRadius = 0.05 * canvasSize;
+  context.font = `${0.08 * canvasSize}px sans-serif`;
+  context.textAlign = 'center';
+  context.textBaseline = 'middle';
+  context.beginPath();
+  const points = previousPoints.filter((point, index, arr) => {
+    return arr.findIndex(x => (x.coordinateX === point.coordinateX && x.coordinateY === point.coordinateY)) === index
+  });
+
+  // draw all points
+  points.forEach(point => {
+    context.fillStyle = fillStyle;
+    context.arc(mapCoordinate(point.coordinateX, canvasSize), mapCoordinate(point.coordinateY, canvasSize), pointRadius, 0, 2 * Math.PI);
+    context.fill();
+  });
+
+  // draw labels over points
+  points.forEach((point, index) => {
+    context.fillStyle = '#ffffff';
+    context.fillText(`${index+1}`, mapCoordinate(point.coordinateX, canvasSize), mapCoordinate(point.coordinateY, canvasSize));
+  });
+}
