+++ frontend/src/services/roomCalibration.tsx
@@ -2,6 +2,7 @@ import { RefObject, useCallback, useContext, useEffect, useState } from 'react';
 import { SocketContext } from './socketProvider';
 import { Acknowledgment } from './acknowledgment';
 import { useAudioMeter, historicVolume, prepareAudioMeter } from './audioMeter';
+import { drawCurrentPosition, drawPoints, mapCoordinate } from './canvasDrawTools';
 
 export type RoomCalibrationRequest = {
   room: {
@@ -42,95 +43,7 @@ export type RoomCalibrationResult = {
   volume: number;
 }
 
-const canvasCordSize = 100;
-const maxCord = 640;
-
-const mapCoordinate = (cord: number): number => {
-  // Remove aliasing using +0.5
-  return Math.round((cord / maxCord) * canvasCordSize) + 0.5;
-}
-
-const getRandomHexColourString = (index: number): string => {
-  const maxAmountOfColours = 16;
-  const hueDelta = Math.trunc(360 / maxAmountOfColours);
-  const h = (index % maxAmountOfColours) * hueDelta;
-  let s = 100;
-  let l = 40;
-
-  // convert hsl to rgb from: https://css-tricks.com/converting-color-spaces-in-javascript/
-  s /= 100;
-  l /= 100;
-
-  let c = (1 - Math.abs(2 * l - 1)) * s,
-      x = c * (1 - Math.abs((h / 60) % 2 - 1)),
-      m = l - c/2,
-      r = 0,
-      g = 0, 
-      b = 0; 
-
-    if (0 <= h && h < 60) {
-      r = c; g = x; b = 0;
-    } else if (60 <= h && h < 120) {
-      r = x; g = c; b = 0;
-    } else if (120 <= h && h < 180) {
-      r = 0; g = c; b = x;
-    } else if (180 <= h && h < 240) {
-      r = 0; g = x; b = c;
-    } else if (240 <= h && h < 300) {
-      r = x; g = 0; b = c;
-    } else if (300 <= h && h < 360) {
-      r = c; g = 0; b = x;
-    }
-    // Having obtained RGB, convert channels to hex
-    let rString = Math.round((r + m) * 255).toString(16);
-    let gString = Math.round((g + m) * 255).toString(16);
-    let bString = Math.round((b + m) * 255).toString(16);
-
-    // Prepend 0s, if necessary
-    if (rString.length === 1)
-      rString = "0" + r;
-    if (gString.length === 1)
-      gString = "0" + g;
-    if (bString.length === 1)
-      bString = "0" + b;
-
-    return "#" + rString + gString + bString;
-}
-
-const drawCurrentPosition = (context: CanvasRenderingContext2D, x: number, y: number) => {
-  const crosshairRadius = 3;
-  context.strokeStyle = '#ff0000';
-  context.beginPath();
-  context.moveTo(x - crosshairRadius, y);
-  context.lineTo(x + crosshairRadius, y);
-  context.moveTo(x, y - crosshairRadius);
-  context.lineTo(x, y + crosshairRadius);
-  context.stroke();
-}
-
-const drawPoints = (context: CanvasRenderingContext2D, previousPoints: RoomCalibrationPoint[], fillStyle: string) => {
-  const pointRadius = 5;
-  context.font = '8px sans-serif';
-  context.textAlign = 'center';
-  context.textBaseline = 'middle';
-  context.beginPath();
-  const points = previousPoints.filter((point, index, arr) => {
-    return arr.findIndex(x => (x.coordinateX === point.coordinateX && x.coordinateY === point.coordinateY)) === index
-  });
-
-  // draw all points
-  points.forEach(point => {
-    context.fillStyle = fillStyle;
-    context.arc(mapCoordinate(point.coordinateX), mapCoordinate(point.coordinateY), pointRadius, 0, 2 * Math.PI);
-    context.fill();
-  });
-
-  // draw labels over points
-  points.forEach((point, index) => {
-    context.fillStyle = '#ffffff';
-    context.fillText(`${index+1}`, mapCoordinate(point.coordinateX), mapCoordinate(point.coordinateY));
-  });
-}
+const canvasSize = 250;
 
 export const useRoomCalibration = (roomId: number, calibrationMapCanvasRef: RefObject<HTMLCanvasElement> | null = null) => {
   const socketRoomName = 'room-calibration';
@@ -148,10 +61,10 @@ export const useRoomCalibration = (roomId: number, calibrationMapCanvasRef: RefO
 
       const context = calibrationMapCanvasRef?.current?.getContext("2d");
       if (context) {
-        context.clearRect(0, 0, canvasCordSize, canvasCordSize);
-        drawPoints(context, roomCalibration.previousPoints, '#555555');
-        drawPoints(context, roomCalibration.currentPoints, '#000000');
-        drawCurrentPosition(context, mapCoordinate(roomCalibration.positionX), mapCoordinate(roomCalibration.positionY));
+        context.clearRect(0, 0, canvasSize, canvasSize);
+        drawPoints(context, canvasSize, roomCalibration.previousPoints, '#555555');
+        drawPoints(context, canvasSize, roomCalibration.currentPoints, '#000000');
+        drawCurrentPosition(context, canvasSize, mapCoordinate(roomCalibration.positionX, canvasSize), mapCoordinate(roomCalibration.positionY, canvasSize));
       }
     }
   }, [roomId, calibrationMapCanvasRef]);
@@ -172,11 +85,11 @@ export const useRoomCalibration = (roomId: number, calibrationMapCanvasRef: RefO
 
   useEffect(() => {
     if (calibrationMapCanvasRef && calibrationMapCanvasRef.current) {
-      calibrationMapCanvasRef.current.width = canvasCordSize;
-      calibrationMapCanvasRef.current.height = canvasCordSize;
+      calibrationMapCanvasRef.current.width = canvasSize;
+      calibrationMapCanvasRef.current.height = canvasSize;
       const context = calibrationMapCanvasRef.current.getContext("2d");
       if (context) {
-        context.clearRect(0, 0, canvasCordSize, canvasCordSize);
+        context.clearRect(0, 0, canvasSize, canvasSize);
       }
     }
   }, [roomCalibration?.calibrating, calibrationMapCanvasRef])
